hw03/requirements.txt
face-recognition>=1.3.0
streamlit>=1.20.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0